Add logic to user edit service, refactor code

+++ desktop/RentDesktop/Models/ITransportRent.cs
@@ -4,6 +4,6 @@
     {
         Transport Transport { get; }
         int Days { get; set; }
-        int TotalPrice { get; }
+        double TotalPrice { get; }
     }
 }
